requirements.txt
configparser==3.3.0r2
httplib2==0.7.7
oauth2==1.5.211
oauthlib==0.6.1
python-twitter==1.3.1
pytz==2012h
redis==2.9.1
requests==2.2.1
requests-oauthlib==0.4.0
simplejson==3.4.0
wsgiref==0.1.2
